MINOR-restructuring: moved NetworkPartitionMetadata in org.apache.kafka.server.util
